datastore: Add a simple multi-source datastore
This datastore encapsulates another one and if a blob with given name
is not present locally, it queries other datastores and updates the
local one if blob can be found there.

Blobs are also periodically fetched from remote storages which is
necessary for updating dynamic links.

The current implementation is heavily unoptimized but can already be
used for testing more distributed scenarios.