dynamiclink: Sign sha512 of the content instead of the content itself
This is mostly to alow streaming the link data. The implementation
of the original ed25519 signature requires two-pass over the dataset
which prevents such streaming operation. With hash, the hash can be
calculated during stream processing.

The result is then equivalent to signing the hash of the message
instead of the message but since the hash corresponds to a single
message only (assuming sha512 is not broken), it does impose the
same cryptographic characteristics.